fix(suite-native): unify receive flow navigation for all use cases
