+++ suite-native/app/src/navigation/RootStackNavigator.tsx
@@ -18,7 +18,7 @@ import { DevUtilsStackNavigator } from '@suite-native/module-dev-utils';
 import { TransactionDetailScreen } from '@suite-native/transactions';
 import { OnboardingStackNavigator as LegacyOnboardingStackNavigator } from '@suite-native/module-onboarding-legacy';
 import { OnboardingStackNavigator } from '@suite-native/module-onboarding';
-import { ReceiveStackNavigator, ReceiveModalScreen } from '@suite-native/receive';
+import { ReceiveStackNavigator } from '@suite-native/receive';
 import { AuthorizeDeviceStackNavigator } from '@suite-native/module-authorize-device';
 import { AddCoinAccountStackNavigator } from '@suite-native/module-add-accounts';
 import { DeviceSettingsStackNavigator } from '@suite-native/module-device-settings';
@@ -101,7 +101,6 @@ export const RootStackNavigator = () => {
                 component={CoinEnablingInitScreen}
                 options={{ animation: 'slide_from_bottom' }}
             />
-            <RootStack.Screen name={RootStackRoutes.ReceiveModal} component={ReceiveModalScreen} />
             <RootStack.Screen
                 name={RootStackRoutes.ReceiveStack}
                 component={ReceiveStackNavigator}
