+++ suite-native/receive/src/index.ts
@@ -1,3 +1,1 @@
 export * from './navigation/ReceiveStackNavigator';
-export * from './screens/ReceiveAccountScreen';
-export * from './screens/ReceiveModalScreen';
