+++ suite-native/receive/tsconfig.json
@@ -27,6 +27,7 @@
         { "path": "../toasts" },
         { "path": "../tokens" },
         { "path": "../../packages/connect" },
+        { "path": "../../packages/env-utils" },
         { "path": "../../packages/styles" },
         { "path": "../../packages/theme" }
     ]
